test(parser): split StatementParser test for PSR-0 and PSR-4
Split StatementParserTest into two separate test classes:
- Psr0StatementParserTest: Tests the legacy PSR-0 implementation (lib/)
- Psr4StatementParserTest: Tests the modern PSR-4 implementation (src/)

The original StatementParserTest.php now contains only a deprecation notice
pointing to the new test files.

Added additional test coverage for PSR-4 StatementParser:
- testParserIteratesOverStatements: Verifies Iterator interface
- testParserHandlesStringInput: Tests string path constructor

Both test files work with PHPUnit 11.5 and 13.0.
